pyproject.toml
[build-system]
requires = ["setuptools"]

[tool.setuptools.packages.find]
# venv folders
exclude = ["bin",  "include",  "lib",  "lib64"]

[project]
name = "sdcard"
description = "USB read-only access to SD card on Android system"
readme = "README.md"
version = "0.0.1"
dependencies = [
"adb_shell>=0.4.4",
"cffi>=1.17.1",
"cryptography>=43.0.1",
"fusepy>=3.0.1",
"libusb1>=3.1.0",
"pyasn1>=0.6.1",
"pycparser>=2.22",
"rsa>=4.9",
]
requires-python = ">=3"

[project.scripts]
sdcard = "sdcard:__main__.main"

[project.urls]
Homepage = "https://github.com/patatetom/sdcard/"
